chore(neovim): disable modulizing
